Migrate to WithCredentialsJSON for GCS
The current WithCredentialsFile approach leaves state in Docker
containers, on disk, etc.

The Google Cloud SDKs support a WithCredentialsJSON option provider that
reads from a byte slice.

I've also changed the credentials format to be []byte and removed the
manual base64 decoding and encoding, as encoding/json wrangles that
automatically.

Support for existing credentials files has been preserved, but they are
no longer written to disk when adding or updating users, and instead go
into the dataprovider as part of the user.