Add first version of an in-memory filesystem implementation
The in-memory filesystem implementation is backed by
spf13/afero. This resulted in a fairly shallow layer
around the methods provided by afero.

So far the basic SFTP commands work, but there are
some fixes to be made, for example around symlinks and
determining paths. These probably also need some tests.