add build flag to entirely disable journald logger
The library used to log to journald go-systemd always creates
a socket on init, no matter if you'll use it or not:

https://github.com/coreos/go-systemd/blob/9ed442a88da2e6230e1d13be1820a5357055aec8/journal/journal_unix.go#L56

Journald logging is "eventually" available only in the subsystem
and only if the user decides to activate.

No matter which part of sftpgo you use, there is always somewhere
where you include the sftpgo/logger, even `sftpgo help` and so
you always get a socket.

This change adds a build flag so sftpgo can be built without
go-systemd/journal dependency and so prevent the creation of
a socket that will not be used.

Signed-off-by: Jeremy Clerc <jeremy@clerc.io>