Add a new config value for default share expirations
This commit adds a new configuration option `share_expiration_days` to
control the date when a newly created share shall expires.

For example, if an administrator declares `share_expiration_days: 30` in
their SFTPGo configuraiton, the expiration date in the "Add a new share"
form would be set to 30 days in the future. The user would still able to
choose a different expiration date (or no expiration at all) in the
form.

Signed-off-by: M. A <mak@nyantec.com>