fix: DataTables type auto does not always detect the correct type for columns.type
By default DataTables detect the column type automatically this does not always work properly as demonstrated in #1628.  In DataTables, columns.type is used for filtering and sorting but also add a class `dt-type-*` for the detected type. This pose an issue if the correct type is not detected because the theme use this class to set the text alignment (eg. numeric type is align to the right)
Fix #1628